move syncTimeWithGoogle out of popup.ts
this prevents a dependency issue in the tests
