+++ src/popup.ts
@@ -18,6 +18,7 @@ import { Notification } from "./store/Notification";
 import { Qr } from "./store/Qr";
 import { Advisor } from "./store/Advisor";
 import { Dropbox, Drive, OneDrive } from "./models/backup";
+import { syncTimeWithGoogle } from "./syncTime";
 
 let LocalStorage: {
   [key: string]: any;
@@ -193,7 +194,7 @@ async function init() {
     { origins: ["https://www.google.com/"] },
     (hasPermission) => {
       if (hasPermission) {
-        syncTimeWithGoogle();
+        syncTimeWithGoogle(LocalStorage);
       }
     }
   );
@@ -337,47 +338,4 @@ async function runScheduledBackup(clientTime: number, instance: Vue) {
     LocalStorage.lastRemindingBackupTime = clientTime;
     chrome.storage.local.set({ LocalStorage });
   }
-}
-
-export function syncTimeWithGoogle() {
-  return new Promise(
-    (resolve: (value: string) => void, reject: (reason: Error) => void) => {
-      try {
-        // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-        // @ts-ignore
-        const xhr = new XMLHttpRequest({ mozAnon: true });
-        xhr.open("HEAD", "https://www.google.com/generate_204");
-        const xhrAbort = setTimeout(() => {
-          xhr.abort();
-          return resolve("updateFailure");
-        }, 5000);
-        xhr.onreadystatechange = () => {
-          if (xhr.readyState === 4) {
-            clearTimeout(xhrAbort);
-            const date = xhr.getResponseHeader("date");
-            if (!date) {
-              return resolve("updateFailure");
-            }
-            const serverTime = new Date(date).getTime();
-            const clientTime = new Date().getTime();
-            const offset = Math.round((serverTime - clientTime) / 1000);
-
-            if (Math.abs(offset) <= 300) {
-              // within 5 minutes
-              LocalStorage.offset = Math.round(
-                (serverTime - clientTime) / 1000
-              );
-              chrome.storage.local.set({ LocalStorage });
-              return resolve("updateSuccess");
-            } else {
-              return resolve("clock_too_far_off");
-            }
-          }
-        };
-        xhr.send();
-      } catch (error) {
-        return reject(error as Error);
-      }
-    }
-  );
 }
